Correction de Dijkstra et ses tests

+++ src/dijkstra.h
@@ -1,8 +1,9 @@
 #ifndef DIJKSTRA_H
 #define DIJKSTRA_H
 #include "graph.h"
-#include "borne.c"
+#include "trip.h"
+#include "borne.h"
 
-Trip *dijkstra(BorneElectrique *,int, float, int, int);
+Trip *dijkstra(BorneElectrique *,int, int, int, int);
 
 #endif // DIJKSTRA_H
